Centralized test runner
Fixes LettError MutatorMath#69 by introducing a new entrypoint for running all the tests. The preferred way of running the tests is now `python setup.py test`, but it is also possible to run them all manually by invoking `python Lib/mutatorMath/test/run.py` or by running each file individually as before. This commit also fixes some issues in the test data, like the lowercase x in the LightWide.ufo master not being point-compatible with the other masters.
