+++ .travis.yml
@@ -8,7 +8,7 @@ install:
   - pip install coveralls
   - pip install -v .
 script:
-  - coverage run --parallel-mode --source=mutatorMath Lib/mutatorMath/test/ufo/test.py
+  - coverage run --parallel-mode --source=mutatorMath Lib/mutatorMath/test/run.py
 after_success:
   - coverage combine
   - coveralls
