+++ Lib/mutatorMath/test/ufo/data/sources/light/LightWide.ufo/glyphs/x.glif
@@ -21,6 +21,7 @@
       <point x="547" y="0" type="line"/>
       <point x="547" y="1" type="line"/>
       <point x="311" y="277" type="line"/>
+      <point x="521" y="527" type="line"/>
     </contour>
   </outline>
 </glyph>
